move toFhir out of person
